*openAL- add resampler quality control cvar *opt light editor rendering
git-svn-id: https://svn.code.sf.net/p/quake2xp/code@1114 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/ref_gl/r_main.c
@@ -1958,7 +1958,7 @@ static float ClampCvar(float min, float max, float value) {
 	return value;
 }
 
-static float ClampCvarInteger(int min, int max, int value) {
+static int ClampCvarInteger(int min, int max, int value) {
 	if (value < min) return min;
 	if (value > max) return max;
 	return value;
